File rearrange to separate classes
